requirements.txt
streamlit>=1.0
pandas>=1.0
scikit-learn>=1.0
joblib>=1.0
numpy>=1.20
tensorflow
pillow

